fasta_print_db_relabel: single uint64_t ordinal to fix macOS ambiguity
Commit 09a01fa1 split fasta_print_db_relabel into two overloads (int and
std::size_t ordinal) to absorb the static_cast at the call sites. That is
ambiguous for a uint64_t argument (mask.cc passes i + 1) on platforms where
uint64_t, size_t and int are all distinct types: macOS, where uint64_t is
unsigned long long but size_t is unsigned long. It built on Linux only
because there size_t and uint64_t are the same type.

Collapse the two overloads into one taking uint64_t (the widest unsigned
type): every caller passes a non-negative 1-based counter as int, size_t or
uint64_t, all of which convert to uint64_t without narrowing or sign change,
so the grouped static_cast and -Wconversion cleanliness are preserved while
the ambiguity is gone. cluster.cc's ordinal becomes uint64_t to match (it is
a pure counter), avoiding an int -> uint64_t sign-conversion warning.

+++ src/cluster.cc
@@ -1384,7 +1384,7 @@ auto cluster(char const * dbname,
   }
 
   int lastcluster = -1;
-  int ordinal = 0;
+  uint64_t ordinal = 0;
 
   for (int i = 0; i < seqcount; i++)
     {
